backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
sqlalchemy==2.0.48
pydantic[email]==2.12.5
python-dotenv==1.2.2
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
email-validator>=2.0.0
slowapi==0.1.9
